+++ src/Views/MainView.xaml.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Input;
 using Lei.Common;
 using Lei.UI;
 using winform = System.Windows.Forms;